+++ changelog.md
@@ -69,6 +69,8 @@
 
 - Fixed buffer overflow bugs in `net`
 
+- Exported `sslHandle` from `net` and `asyncnet`.
+
 - Added `sections` iterator in `parsecfg`.
 
 - Make custom op in macros.quote work for all statements.
@@ -294,6 +296,8 @@
 
 - Added `genasts.genAst` that avoids the problems inherent with `quote do` and can
   be used as a replacement.
+
+- Added `copyWithin` [for `seq` and `array` for JavaScript targets](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/copyWithin).
 
 
 ## Language changes